+++ src/lib/Scenes/Artwork/Components/ArtworkTombstone.tsx
@@ -1,9 +1,10 @@
-import { Box, Flex, Sans, Serif } from "@artsy/palette"
+import { Box, Flex, Serif } from "@artsy/palette"
 import { ArtworkTombstone_artwork } from "__generated__/ArtworkTombstone_artwork.graphql"
 import SwitchBoard from "lib/NativeModules/SwitchBoard"
 import React from "react"
 import { NativeModules, TouchableWithoutFeedback } from "react-native"
 import { createFragmentContainer, graphql } from "react-relay"
+import { FollowArtistButtonFragmentContainer as FollowArtistButton } from "./FollowArtistButton"
 
 const Constants = NativeModules.ARCocoaConstantsModule
 
@@ -36,12 +37,7 @@ export class ArtworkTombstone extends React.Component<ArtworkTombstoneProps, Art
       <React.Fragment>
         <Flex flexDirection="row">
           {this.renderArtistName(artist.name, artist.href)}
-          <Sans color="black60" size="6" mx={1}>
-            &middot;
-          </Sans>
-          <Sans color="black60" size="4">
-            Follow
-          </Sans>
+          <FollowArtistButton artist={artist} />
         </Flex>
       </React.Fragment>
     )
@@ -137,6 +133,7 @@ export const ArtworkTombstoneFragmentContainer = createFragmentContainer(Artwork
       artists {
         name
         href
+        ...FollowArtistButton_artist
       }
       dimensions {
         in