+++ src/lib/Scenes/Artwork/Components/__stories__/ArtworkTombstone.story.tsx
@@ -1,108 +1,44 @@
 import { storiesOf } from "@storybook/react-native"
+import { ArtworkTombstone_artwork } from "__generated__/ArtworkTombstone_artwork.graphql"
+import { MockRelayRenderer } from "lib/tests/MockRelayRenderer"
 import React from "react"
-import { ArtworkTombstone } from "../ArtworkTombstone"
-
-storiesOf("Artwork/Components")
-  .add("Tombstone with 1 artist", () => {
-    return (
-      <ArtworkTombstone
-        artwork={
-          {
-            title: "Hello im a title",
-            medium: "Painting",
-            date: "1992",
-            artists: [{ name: "Andy Warhol", href: "/artist/pablo-picasso" }],
-            cultural_maker: null,
-            dimensions: {
-              in: "15 × 20 in",
-              cm: "38.1 × 50.8 cm",
-            },
-            edition_of: "Edition 100/200",
-            attribution_class: {
-              short_description: "This is an edition of something",
-            },
-          } as any
+import { graphql } from "react-relay"
+import { ArtworkTombstoneFragmentContainer } from "../ArtworkTombstone"
+const MockArtworkTombstone = ({ artwork }: { artwork: ArtworkTombstone_artwork }) => {
+  return (
+    <MockRelayRenderer
+      Component={ArtworkTombstoneFragmentContainer}
+      mockData={{ artwork }}
+      query={graphql`
+        query ArtworkTombstoneStoryQuery {
+          artwork(id: "unused") {
+            ...ArtworkTombstone_artwork
+          }
         }
-      />
-    )
-  })
-  .add("Tombstone with many artists", () => {
-    return (
-      <ArtworkTombstone
-        artwork={
-          {
-            title: "Hello im a title",
-            medium: "Painting",
-            date: "1992",
-            artists: [
-              { name: "Andy Warhol", href: "/artist/pablo-picasso" },
-              { name: "Alex Katz", href: "/artist/alex-katz" },
-              { name: "Pablo Picasso", href: "/artist/pablo-picasso" },
-              { name: "Banksy", href: "/artist/banksy" },
-              { name: "Robert Trundelbed Bananahammock Verylongname", href: "/artist/banksy" },
-              { name: "Barbara Kruger", href: "/artist/barbara-kruger" },
-            ],
-            cultural_maker: null,
-            dimensions: {
-              in: "15 × 20 in",
-              cm: "38.1 × 50.8 cm",
-            },
-            edition_of: "Edition 100/200",
-            attribution_class: {
-              short_description: "This is an edition of something",
-            },
-          } as any
-        }
-      />
-    )
-  })
-  .add("Tombstone with 3 artists", () => {
-    return (
-      <ArtworkTombstone
-        artwork={
-          {
-            title: "Hello im a title",
-            medium: "Painting",
-            date: "1992",
-            artists: [
-              { name: "Andy Warhol", href: "/artist/pablo-picasso" },
-              { name: "Alex Katz", href: "/artist/alex-katz" },
-              { name: "Pablo Picasso", href: "/artist/pablo-picasso" },
-            ],
-            cultural_maker: null,
-            dimensions: {
-              in: "15 × 20 in",
-              cm: "38.1 × 50.8 cm",
-            },
-            edition_of: "Edition 100/200",
-            attribution_class: {
-              short_description: "This is an edition of something",
-            },
-          } as any
-        }
-      />
-    )
-  })
-  .add("tombstone with cultural maker", () => {
-    return (
-      <ArtworkTombstone
-        artwork={
-          {
-            title: "Hello im a title",
-            medium: "Painting",
-            date: "1992",
-            artists: [],
-            cultural_maker: "some guy idk",
-            dimensions: {
-              in: "15 × 20 in",
-              cm: "38.1 × 50.8 cm",
-            },
-            edition_of: "Edition 100/200",
-            attribution_class: {
-              short_description: "This is an edition of something",
-            },
-          } as any
-        }
-      />
-    )
-  })
+      `}
+    />
+  )
+}
+const artworkTombstoneArtwork = {
+  title: "Hello im a title",
+  medium: "Painting",
+  date: "1992",
+  artists: [
+    {
+      name: "Andy Warhol",
+      href: "/artist/andy-warhol",
+      id: "1234",
+      gravityID: "andy-warhol",
+      is_followed: false,
+      " $fragmentRefs": null,
+    },
+  ],
+  cultural_maker: null,
+  dimensions: { in: "15 × 20 in", cm: "38.1 × 50.8 cm" },
+  edition_of: "Edition 100/200",
+  attribution_class: { short_description: "This is an edition of something" },
+  " $refType": null,
+}
+storiesOf("Artwork/Components").add("Tombstone", () => {
+  return <MockArtworkTombstone artwork={artworkTombstoneArtwork} />
+})